refactor[#33]: AlbumFinder AlbumManger로 리팩토링
