+++ src/main/kotlin/com/vibetrip/vibetripserver/album/business/AlbumService.kt
@@ -3,7 +3,6 @@ package com.vibetrip.vibetripserver.album.business
 import com.vibetrip.vibetripserver.album.domain.NewAlbum
 import com.vibetrip.vibetripserver.album.implement.AiProcessor
 import com.vibetrip.vibetripserver.album.implement.AlbumCoverImageProcessor
-import com.vibetrip.vibetripserver.album.implement.AlbumFinder
 import com.vibetrip.vibetripserver.album.implement.AlbumManager
 import com.vibetrip.vibetripserver.album.presentation.dto.response.AlbumCreateResponse
 import com.vibetrip.vibetripserver.support.paging.Cursorable
@@ -15,7 +14,6 @@ import org.springframework.web.multipart.MultipartFile
 class AlbumService(
     private val albumManager: AlbumManager,
     private val aiProcessor: AiProcessor,
-    private val albumFinder: AlbumFinder,
     private val albumCoverImageProcessor: AlbumCoverImageProcessor,
 ) {
     @Transactional
@@ -35,7 +33,7 @@ class AlbumService(
     fun findAlbums(
         memberKey: String,
         cursorable: Cursorable<Long>,
-    ) = albumFinder.findAllByMemberKey(memberKey, cursorable)
+    ) = albumManager.find(memberKey, cursorable)
 
-    fun countAlbums(memberKey: String): Long = albumFinder.countByMemberKey(memberKey)
+    fun countAlbums(memberKey: String): Long = albumManager.count(memberKey)
 }
